+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+### Added
+
+- Update to v4 of app library and skip if app dependencies are not ready to
+allow app CR creation.
 
 ## [Unreleased]